+++ aries-core/src/js/components/layouts/index.js
@@ -1,1 +1,3 @@
+export * from './ButtonGroup';
+export * from './Layer';
 export * from './ModalDialog';
