+++ aries-site/src/examples/templates/feedback/FeedbackAnatomy.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Box, Button, Diagram, Grid, Header, Heading, Stack } from 'grommet';
-import { ButtonGroup } from '../ButtonGroup';
+import { ButtonGroup } from 'aries-core';
 import { Annotation } from '../../../layouts';
 import { connection } from '../../../utils';
