+++ aries-site/src/examples/templates/feedback/ModalTaskFlowExample/TaskFlowFeedback.js
@@ -9,7 +9,7 @@ import {
   Text,
   TextArea,
 } from 'grommet';
-import { ButtonGroup } from '../../ButtonGroup';
+import { ButtonGroup } from 'aries-core/src/js/components/layouts/ButtonGroup';
 
 // These credentials will be supplied when working with HPE GreenLake Experience Management COE team
 // and should likely be stored in environment variables.
