+++ aries-site/src/examples/templates/forms/managing-child-objects/CreateCluster.js
@@ -11,7 +11,7 @@ import {
   Text,
   TextInput,
 } from 'grommet';
-import { ButtonGroup } from '../../ButtonGroup';
+import { ButtonGroup } from 'aries-core';
 import { FormChildObjects } from '../../FormChildObject';
 
 const hostTemplate = {
