+++ aries-site/src/examples/templates/forms/managing-child-objects/CreateRole.js
@@ -9,7 +9,7 @@ import {
   Select,
   TextInput,
 } from 'grommet';
-import { ButtonGroup } from '../../ButtonGroup';
+import { ButtonGroup } from 'aries-core';
 import { FormChildObjects } from '../../FormChildObject';
 
 const permissionTemplate = {
